requirements.txt
torch
torchvision
tokenizers
faiss-cpu
streamlit
ftfy
regex
tqdm
click
